move and split defaults methods,
- expose main supplyLayoutDefaults only on Axes
