+++ src/plots/cartesian/axes.js
@@ -16,6 +16,7 @@ var isNumeric = require('fast-isnumeric');
 var axes = module.exports = {};
 
 axes.layoutAttributes = require('./layout_attributes');
+axes.supplyLayoutDefaults = require('./layout_defaults');
 
 axes.setConvert = require('./set_convert');
 
@@ -26,321 +27,7 @@ axes.list = utils.list;
 axes.listIds = utils.listIds;
 axes.getFromId = utils.getFromId;
 axes.getFromTrace = utils.getFromTrace;
-axes.supplyLayoutDefaults = function(layoutIn, layoutOut, fullData) {
-    // get the full list of axes already defined
-    var layoutKeys = Object.keys(layoutIn),
-        xaList = [],
-        yaList = [],
-        outerTicks = {},
-        noGrids = {},
-        i;
 
-    for(i = 0; i < layoutKeys.length; i++) {
-        var key = layoutKeys[i];
-        if(xAxisMatch.test(key)) xaList.push(key);
-        else if(yAxisMatch.test(key)) yaList.push(key);
-    }
-
-    for(i = 0; i < fullData.length; i++) {
-        var trace = fullData[i],
-            xaName = axes.id2name(trace.xaxis),
-            yaName = axes.id2name(trace.yaxis);
-
-        // add axes implied by traces
-        if(xaName && xaList.indexOf(xaName)===-1) xaList.push(xaName);
-        if(yaName && yaList.indexOf(yaName)===-1) yaList.push(yaName);
-
-        // check for default formatting tweaks
-        if(Plotly.Plots.traceIs(trace, '2dMap')) {
-            outerTicks[xaName] = true;
-            outerTicks[yaName] = true;
-        }
-
-        if(Plotly.Plots.traceIs(trace, 'oriented')) {
-            var positionAxis = trace.orientation==='h' ? yaName : xaName;
-            noGrids[positionAxis] = true;
-        }
-    }
-
-    function axSort(a,b) {
-        var aNum = Number(a.substr(5)||1),
-            bNum = Number(b.substr(5)||1);
-        return aNum - bNum;
-    }
-
-    if(layoutOut._hasCartesian || layoutOut._hasGL2D || !fullData.length) {
-        // make sure there's at least one of each and lists are sorted
-        if(!xaList.length) xaList = ['xaxis'];
-        else xaList.sort(axSort);
-
-        if(!yaList.length) yaList = ['yaxis'];
-        else yaList.sort(axSort);
-    }
-
-    xaList.concat(yaList).forEach(function(axName){
-        var axLetter = axName.charAt(0),
-            axLayoutIn = layoutIn[axName] || {},
-            axLayoutOut = {},
-            defaultOptions = {
-                letter: axLetter,
-                font: layoutOut.font,
-                outerTicks: outerTicks[axName],
-                showGrid: !noGrids[axName],
-                name: axName,
-                data: fullData
-            },
-            positioningOptions = {
-                letter: axLetter,
-                counterAxes: {x: yaList, y: xaList}[axLetter].map(axes.name2id),
-                overlayableAxes: {x: xaList, y: yaList}[axLetter].filter(function(axName2){
-                    return axName2!==axName && !(layoutIn[axName2]||{}).overlaying;
-                }).map(axes.name2id)
-            };
-
-        function coerce(attr, dflt) {
-            return Plotly.Lib.coerce(axLayoutIn, axLayoutOut,
-                                     axes.layoutAttributes,
-                                     attr, dflt);
-        }
-
-        axes.handleAxisDefaults(axLayoutIn, axLayoutOut, coerce, defaultOptions);
-        axes.handleAxisPositioningDefaults(axLayoutIn, axLayoutOut, coerce, positioningOptions);
-        layoutOut[axName] = axLayoutOut;
-
-        // so we don't have to repeat autotype unnecessarily,
-        // copy an autotype back to layoutIn
-        if(!layoutIn[axName] && axLayoutIn.type!=='-') {
-            layoutIn[axName] = {type: axLayoutIn.type};
-        }
-
-    });
-
-    // plot_bgcolor only makes sense if there's a (2D) plot!
-    // TODO: bgcolor for each subplot, to inherit from the main one
-    if(xaList.length && yaList.length) {
-        Plotly.Lib.coerce(layoutIn, layoutOut,
-            Plotly.Plots.layoutAttributes, 'plot_bgcolor');
-    }
-};
-
-/**
- * options: object containing:
- *      letter: 'x' or 'y'
- *      title: name of the axis (ie 'Colorbar') to go in default title
- *      name: axis object name (ie 'xaxis') if one should be stored
- *      font: the default font to inherit
- *      outerTicks: boolean, should ticks default to outside?
- *      showGrid: boolean, should gridlines be shown by default?
- *      noHover: boolean, this axis doesn't support hover effects?
- *      data: the plot data to use in choosing auto type
- */
-axes.handleAxisDefaults = function(containerIn, containerOut, coerce, options) {
-    var letter = options.letter,
-        font = options.font || {},
-        defaultTitle = 'Click to enter ' +
-            (options.title || (letter.toUpperCase() + ' axis')) +
-            ' title';
-
-    // set up some private properties
-    if(options.name) {
-        containerOut._name = options.name;
-        containerOut._id = axes.name2id(options.name);
-    }
-
-    // now figure out type and do some more initialization
-    var axType = coerce('type');
-    if(axType==='-') {
-        setAutoType(containerOut, options.data);
-
-        if(containerOut.type==='-') {
-            containerOut.type = 'linear';
-        }
-        else {
-            // copy autoType back to input axis
-            // note that if this object didn't exist
-            // in the input layout, we have to put it in
-            // this happens in the main supplyDefaults function
-            axType = containerIn.type = containerOut.type;
-        }
-    }
-    axes.setConvert(containerOut);
-
-    coerce('title', defaultTitle);
-    Plotly.Lib.coerceFont(coerce, 'titlefont', {
-        family: font.family,
-        size: Math.round(font.size * 1.2),
-        color: font.color
-    });
-
-    var validRange = (containerIn.range||[]).length===2 &&
-            isNumeric(containerIn.range[0]) &&
-            isNumeric(containerIn.range[1]),
-        autoRange = coerce('autorange', !validRange);
-
-    if(autoRange) coerce('rangemode');
-    var range = coerce('range', [-1, letter==='x' ? 6 : 4]);
-    if(range[0] === range[1]) {
-        containerOut.range = [range[0] - 1, range[0] + 1];
-    }
-    Plotly.Lib.noneOrAll(containerIn.range, containerOut.range, [0, 1]);
-
-    coerce('fixedrange');
-
-    axes.handleTickValueDefaults(containerIn, containerOut, coerce, axType);
-
-    axes.handleTickDefaults(containerIn, containerOut, coerce, axType, options);
-
-
-
-    var lineColor = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'linecolor'),
-        lineWidth = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'linewidth'),
-        showLine = coerce('showline', !!lineColor || !!lineWidth);
-
-    if(!showLine) {
-        delete containerOut.linecolor;
-        delete containerOut.linewidth;
-    }
-
-    if(showLine || containerOut.ticks) coerce('mirror');
-
-    var gridColor = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'gridcolor'),
-        gridWidth = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'gridwidth'),
-        showGridLines = coerce('showgrid', options.showGrid || !!gridColor || !!gridWidth);
-
-    if(!showGridLines) {
-        delete containerOut.gridcolor;
-        delete containerOut.gridwidth;
-    }
-
-    var zeroLineColor = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'zerolinecolor'),
-        zeroLineWidth = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'zerolinewidth'),
-        showZeroLine = coerce('zeroline', options.showGrid || !!zeroLineColor || !!zeroLineWidth);
-
-    if(!showZeroLine) {
-        delete containerOut.zerolinecolor;
-        delete containerOut.zerolinewidth;
-    }
-
-    return containerOut;
-};
-
-/**
- * options: inherits font, outerTicks, noHover from axes.handleAxisDefaults
- */
-axes.handleTickDefaults = function(containerIn, containerOut, coerce, axType, options) {
-    var tickLen = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'ticklen'),
-        tickWidth = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'tickwidth'),
-        tickColor = Plotly.Lib.coerce2(containerIn, containerOut, axes.layoutAttributes, 'tickcolor'),
-        showTicks = coerce('ticks', (options.outerTicks || tickLen || tickWidth || tickColor) ? 'outside' : '');
-    if(!showTicks) {
-        delete containerOut.ticklen;
-        delete containerOut.tickwidth;
-        delete containerOut.tickcolor;
-    }
-
-    var showTickLabels = coerce('showticklabels');
-    if(showTickLabels) {
-        Plotly.Lib.coerceFont(coerce, 'tickfont', options.font || {});
-        coerce('tickangle');
-
-        var showAttrDflt = axes.getShowAttrDflt(containerIn);
-
-        if(axType !== 'category') {
-            var tickFormat = coerce('tickformat');
-            if(!options.noHover) coerce('hoverformat');
-
-            if(!tickFormat && axType !== 'date') {
-                coerce('showexponent', showAttrDflt);
-                coerce('exponentformat');
-            }
-        }
-
-        var tickPrefix = coerce('tickprefix');
-        if(tickPrefix) coerce('showtickprefix', showAttrDflt);
-
-        var tickSuffix = coerce('ticksuffix');
-        if(tickSuffix) coerce('showticksuffix', showAttrDflt);
-    }
-};
-
-axes.handleTickValueDefaults = function(containerIn, containerOut, coerce, axType) {
-    var tickmodeDefault = 'auto';
-
-    if(containerIn.tickmode === 'array' &&
-            (axType === 'log' || axType === 'date')) {
-        containerIn.tickmode = 'auto';
-    }
-
-    if(Array.isArray(containerIn.tickvals)) tickmodeDefault = 'array';
-    else if(containerIn.dtick && isNumeric(containerIn.dtick)) {
-        tickmodeDefault = 'linear';
-    }
-    var tickmode = coerce('tickmode', tickmodeDefault);
-
-    if(tickmode === 'auto') coerce('nticks');
-    else if(tickmode === 'linear') {
-        coerce('tick0');
-        coerce('dtick');
-    }
-    else {
-        var tickvals = coerce('tickvals');
-        if(tickvals === undefined) containerOut.tickmode = 'auto';
-        else coerce('ticktext');
-    }
-};
-
-axes.handleAxisPositioningDefaults = function(containerIn, containerOut, coerce, options) {
-    var counterAxes = options.counterAxes || [],
-        overlayableAxes = options.overlayableAxes || [],
-        letter = options.letter;
-
-    var anchor = Plotly.Lib.coerce(containerIn, containerOut,
-        {
-            anchor: {
-                valType:'enumerated',
-                values: ['free'].concat(counterAxes),
-                dflt: isNumeric(containerIn.position) ? 'free' :
-                    (counterAxes[0] || 'free')
-            }
-        },
-        'anchor');
-
-    if(anchor==='free') coerce('position');
-
-    Plotly.Lib.coerce(containerIn, containerOut,
-        {
-            side: {
-                valType: 'enumerated',
-                values: letter==='x' ? ['bottom', 'top'] : ['left', 'right'],
-                dflt: letter==='x' ? 'bottom' : 'left'
-            }
-        },
-        'side');
-
-    var overlaying = false;
-    if(overlayableAxes.length) {
-        overlaying = Plotly.Lib.coerce(containerIn, containerOut, {
-            overlaying: {
-                valType: 'enumerated',
-                values: [false].concat(overlayableAxes),
-                dflt: false
-            }
-        },
-        'overlaying');
-    }
-
-    if(!overlaying) {
-        // TODO: right now I'm copying this domain over to overlaying axes
-        // in ax.setscale()... but this means we still need (imperfect) logic
-        // in the axes popover to hide domain for the overlaying axis.
-        // perhaps I should make a private version _domain that all axes get???
-        var domain = coerce('domain');
-        if(domain[0] > domain[1] - 0.01) containerOut.domain = [0,1];
-        Plotly.Lib.noneOrAll(containerIn.domain, containerOut.domain, [0, 1]);
-    }
-
-    return containerOut;
-};
 
 // find the list of possible axes to reference with an xref or yref attribute
 // and coerce it to that list
@@ -378,160 +65,6 @@ axes.clearTypes = function(gd, traces) {
 // this can also be used as the id for default counter axis
 axes.counterLetter = function(id) {
     return {x:'y',y:'x'}[id.charAt(0)];
-};
-
-function setAutoType(ax, data){
-    // new logic: let people specify any type they want,
-    // only autotype if type is '-'
-    if(ax.type!=='-') return;
-
-    var id = ax._id,
-        axLetter = id.charAt(0);
-
-    // support 3d
-    if(id.indexOf('scene') !== -1) id = axLetter;
-
-    var d0 = getFirstNonEmptyTrace(data, id, axLetter);
-    if(!d0) return;
-
-    // first check for histograms, as the count direction
-    // should always default to a linear axis
-    if(d0.type==='histogram' &&
-            axLetter==={v:'y', h:'x'}[d0.orientation || 'v']) {
-        ax.type='linear';
-        return;
-    }
-
-    // check all boxes on this x axis to see
-    // if they're dates, numbers, or categories
-    if(isBoxWithoutPositionCoords(d0, axLetter)) {
-        var posLetter = getBoxPosLetter(d0),
-            boxPositions = [],
-            trace;
-
-        for(var i = 0; i < data.length; i++) {
-            trace = data[i];
-            if(!Plotly.Plots.traceIs(trace, 'box') ||
-               (trace[axLetter + 'axis'] || axLetter) !== id) continue;
-
-            if(trace[posLetter] !== undefined) boxPositions.push(trace[posLetter][0]);
-            else if(trace.name !== undefined) boxPositions.push(trace.name);
-            else boxPositions.push('text');
-        }
-
-        ax.type = axes.autoType(boxPositions);
-    }
-    else {
-        ax.type = axes.autoType(d0[axLetter] || [d0[axLetter+'0']]);
-    }
-}
-
-function getBoxPosLetter(trace) {
-    return {v:'x', h:'y'}[trace.orientation || 'v'];
-}
-
-function isBoxWithoutPositionCoords(trace, axLetter) {
-    var posLetter = getBoxPosLetter(trace);
-    return Plotly.Plots.traceIs(trace, 'box') && axLetter===posLetter &&
-            trace[posLetter]===undefined && trace[posLetter + '0']===undefined;
-}
-
-function getFirstNonEmptyTrace(data, id, axLetter) {
-    var trace;
-
-    for(var i = 0; i < data.length; i++) {
-        trace = data[i];
-
-        if((trace[axLetter + 'axis'] || axLetter) === id) {
-            if(isBoxWithoutPositionCoords(trace, axLetter)) {
-                return trace;
-            }
-            else if((trace[axLetter] || []).length || trace[axLetter + '0']) {
-                return trace;
-            }
-        }
-    }
-}
-
-axes.autoType = function(array) {
-    if(axes.moreDates(array)) return 'date';
-    if(axes.category(array)) return 'category';
-    if(linearOK(array)) return 'linear';
-    else return '-';
-};
-
-/*
- * Attributes 'showexponent', 'showtickprefix' and 'showticksuffix'
- * share values.
- *
- * If only 1 attribute is set,
- * the remaining attributes inherit that value.
- *
- * If 2 attributes are set to the same value,
- * the remaining attribute inherits that value.
- *
- * If 2 attributes are set to different values,
- * the remaining is set to its dflt value.
- *
- */
-axes.getShowAttrDflt = function getShowAttrDflt(containerIn) {
-    var showAttrsAll = ['showexponent',
-                        'showtickprefix',
-                        'showticksuffix'],
-        showAttrs = showAttrsAll.filter(function(a){
-            return containerIn[a]!==undefined;
-        }),
-        sameVal = function(a){
-            return containerIn[a]===containerIn[showAttrs[0]];
-        };
-    if (showAttrs.every(sameVal) || showAttrs.length===1) {
-        return containerIn[showAttrs[0]];
-    }
-};
-
-// is there at least one number in array? If not, we should leave
-// ax.type empty so it can be autoset later
-function linearOK(array) {
-    if(!array) return false;
-    for(var i = 0; i < array.length; i++) {
-        if(isNumeric(array[i])) return true;
-    }
-    return false;
-}
-
-// does the array a have mostly dates rather than numbers?
-// note: some values can be neither (such as blanks, text)
-// 2- or 4-digit integers can be both, so require twice as many
-// dates as non-dates, to exclude cases with mostly 2 & 4 digit
-// numbers and a few dates
-axes.moreDates = function(a) {
-    var dcnt=0, ncnt=0,
-        // test at most 1000 points, evenly spaced
-        inc = Math.max(1,(a.length-1)/1000),
-        ai;
-    for(var i=0; i<a.length; i+=inc) {
-        ai = a[Math.round(i)];
-        if(Plotly.Lib.isDateTime(ai)) dcnt+=1;
-        if(isNumeric(ai)) ncnt+=1;
-    }
-    return (dcnt>ncnt*2);
-};
-
-// are the (x,y)-values in td.data mostly text?
-// require twice as many categories as numbers
-axes.category = function(a) {
-    // test at most 1000 points
-    var inc = Math.max(1, (a.length - 1) / 1000),
-        curvenums = 0,
-        curvecats = 0,
-        ai;
-
-    for(var i = 0; i < a.length; i += inc) {
-        ai = axes.cleanDatum(a[Math.round(i)]);
-        if(isNumeric(ai)) curvenums++;
-        else if(typeof ai === 'string' && ai !== '' && ai !== 'None') curvecats++;
-    }
-    return curvecats > curvenums * 2;
 };
 
 // incorporate a new minimum difference and first tick into
